services/karton-kernelsense/requirements.txt
karton-core>=5.0.0
jsonschema>=4.17.0
anthropic>=0.40.0
redis>=5.0.0
chromadb>=0.4.0
